+++ src/main/java/com/yeoljeong/tripmate/notification/infrastructure/persistence/jpa/NotificationSendOutboxJpaRepository.java
@@ -1,5 +1,6 @@
-package com.yeoljeong.tripmate.notification.infrastructure.persistence;
+package com.yeoljeong.tripmate.notification.infrastructure.persistence.jpa;
 
+import com.yeoljeong.tripmate.notification.domain.model.NotificationSendOutbox;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
